refactor(vta-sdk): split VtaClient into per-domain impl blocks
vta-sdk/src/client.rs (~1,858 lines, 60+ methods) is now a module
directory. Each domain (auth/keys/acl/contexts/webvh/services/bootstrap/
did_templates/backup/audit/secrets/vta-management) lives in its own file
with its own `impl VtaClient { ... }` block. The struct definition,
constructor, and shared HTTP plumbing (rpc, ensure_token_valid,
handle_response, etc.) stay in client/mod.rs as `pub(super)` so per-
domain modules reach them through `super::*`. No public-API change.

Also fixes two pre-existing build errors in
vta-sdk/tests/provision_client_e2e.rs that c5e1ba4 missed when adding
the `create_context` field to `ProvisionIntegrationRequest` and
`context_created` to `ProvisionSummary` — the test still constructed
both literals with the old shape, breaking `cargo test -p vta-sdk
--all-features` until now.

Refs #8 in the workspace audit.
