+++ vta-sdk/tests/provision_client_e2e.rs
@@ -513,6 +513,7 @@ async fn admin_rotated_via_rest_round_trip() {
         context: "ctx-1".into(),
         assertion: None,
         vc_validity_seconds: None,
+        create_context: false,
     };
     let response = client
         .provision_integration(req)
@@ -618,6 +619,7 @@ async fn admin_rotated_didcomm_response_decoder_extracts_rotated_credentials() {
             secret_count: 1,
             output_count: 0,
             webvh_server_id: None,
+            context_created: false,
         },
     };
